Updated api/quotes.js
Updated all 1700+ lines of the file to use the new
quote format. This took a week and a half to
complete, but it was worth it.
Will revisit this to optimize the code further in
the future by utlizing the map function to
reduce the number of lines in the file.